docker/api/requirements.txt
fastapi[all]
uvicorn
SQLAlchemy
psycopg2-binary
python-dotenv